pyproject.toml
[project]
name = "ASR-bot"
version = "0.1.0"
authors = [
    {name = "Ailurus1"},
]
description = ""

dependencies = [
    "pre-commit",
    "pytest"
]

[project.optional-dependencies]
bot = [
    "python-telegram-bot",
    "moviepy ==1.0.3",
    "requests"
]
asr-service = [
    "fastapi",
    "numpy <2.0.0",
    "uvicorn",
    "torch ==2.5.1",
    "transformers",
    "torchaudio",
    "soundfile",
    "python-multipart",
    "peft",
    "setuptools",
    "pyannote.audio",
    "pyannote.core"
]
train = [
    "numpy <2.0.0",
    "torch",
    "torchaudio ==2.5.1",
    "transformers",
    "polars",
    "jiwer",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu"},
]
torchaudio = [
    { index = "pytorch-cpu"},
]

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["inference_server", "modeling", "bot"]

[tool.poetry.dependencies]
pyannote-audio = "*"
pyannote-core = "*"
